update lsft_printf && get_next_line
