+++ Makefile
@@ -9,25 +9,23 @@ RM = rm -f
 
 
 SRCS = ft_isalpha.c ft_isdigit.c ft_isalnum.c ft_isascii.c \
-       ft_isprint.c ft_strlen.c ft_memset.c ft_bzero.c \
-       ft_memcpy.c ft_memmove.c ft_strlcpy.c ft_strlcat.c \
-       ft_calloc.c ft_strdup.c ft_toupper.c ft_tolower.c \
-       ft_strchr.c ft_strncmp.c ft_memchr.c ft_memcmp.c \
-       ft_strnstr.c ft_atoi.c ft_strrchr.c ft_substr.c \
-       ft_strjoin.c ft_strtrim.c ft_split.c ft_itoa.c \
-       ft_strmapi.c ft_striteri.c ft_putchar_fd.c \
-       ft_putstr_fd.c ft_putendl_fd.c ft_putnbr_fd.c \
+			 ft_isprint.c ft_strlen.c ft_memset.c ft_bzero.c \
+			 ft_memcpy.c ft_memmove.c ft_strlcpy.c ft_strlcat.c \
+			 ft_calloc.c ft_strdup.c ft_toupper.c ft_tolower.c \
+			 ft_strchr.c ft_strncmp.c ft_memchr.c ft_memcmp.c \
+			 ft_strnstr.c ft_atoi.c ft_strrchr.c ft_substr.c \
+			 ft_strjoin.c ft_strtrim.c ft_split.c ft_itoa.c \
+			 ft_strmapi.c ft_striteri.c ft_putchar_fd.c \
+			 ft_putstr_fd.c ft_putendl_fd.c ft_putnbr_fd.c \
 			 ft_isspace.c ft_free_str_lst.c ft_atol.c \
-			 ft_is_str_chars_same.c \
 			 ft_lstnew_bonus.c ft_lstadd_front_bonus.c \
 			 ft_lstsize_bonus.c ft_lstlast_bonus.c \
 			 ft_lstadd_back_bonus.c ft_lstdelone_bonus.c \
 			 ft_lstclear_bonus.c ft_lstiter_bonus.c \
 			 ft_lstmap_bonus.c \
-			 ft_printf.c ft_print_char.c ft_print_str.c \
-			 ft_print_ptr.c ft_print_nbr.c \
-			 ft_print_unsigned_nbr.c \
-			 ft_print_hexa.c ft_get_next_line.c \
+			 ft_printf.c ft_printf_utils.c \
+			 ft_get_next_line.c \
+			 ft_is_str_chars_same.c \
 			 ft_strndup.c \
 			 ft_malloc.c
