--- ft_print_str.c
@@ -1,31 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_print_str.c                                     :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: znajdaou <znajdaou@student.1337.ma>        +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/12/02 15:38:43 by znajdaou          #+#    #+#             */
-/*   Updated: 2024/12/02 15:44:28 by znajdaou         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "ft_printf.h"
-
-int	ft_print_str(char *str)
-{
-	int	out_len;
-
-	out_len = 0;
-	if (!str)
-	{
-		write(1, "(null)", 6);
-		return (6);
-	}
-	while (*str)
-	{
-		out_len += ft_print_char(*str);
-		str++;
-	}
-	return (out_len);
-}
