+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: znajdaou <znajdaou@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/25 14:42:48 by znajdaou          #+#    #+#             */
-/*   Updated: 2024/12/31 11:55:04 by znajdaou         ###   ########.fr       */
+/*   Updated: 2024/12/31 12:57:51 by znajdaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,8 @@ char	*ft_strchr(const char *s, int c)
 {
 	size_t	i;
 
+  if (!s)
+    return (NULL);
 	i = 0;
 	while (s[i] != (char)c && s[i])
 		i++;
